% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{agg_one}
\alias{agg_one}
\title{Aggregate individual-level CP to One CP}
\usage{
agg_one(
  data,
  ytildename,
  iname,
  tname,
  bname,
  aname,
  kname,
  a_min = 25,
  a_max = 34,
  k_min = 0,
  k_max = 5
)
}
\arguments{
\item{data}{The dataframe containing the individual-level CP (as \code{ytildename})}

\item{ytildename}{Individual-level CP estimated by \code{est_indcp}}

\item{iname}{Individual identifier}

\item{tname}{Time variable}

\item{bname}{Birth year variable}

\item{aname}{Age at treatment variable}

\item{kname}{Relative time to treatment}

\item{a_min}{Minimum age at treatment for aggregation}

\item{a_max}{Maximum age at treatment for aggregation}

\item{k_min}{Minimum relative time to treatment for aggregation}

\item{k_max}{Maximum relative time to treatment for aggregation}
}
\value{
A dataframe with the aggregated CP
}
\description{
Aggregate individual-level CP to One CP
}
