% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_indcp}
\alias{aggregate_indcp}
\title{Aggregate the mean and variance of the estimated individual-level child penalties}
\usage{
aggregate_indcp(object, agg = "full", na.rm = TRUE)
}
\arguments{
\item{object}{\code{indcp} object}

\item{agg}{Aggregation method.
One of \code{c("full", "event", "event_age")} and the default is \code{full}.
If \code{by} is provided in the model,
all the options will separately aggregate by its group.
The \code{event} option aggregates by the group of the event timing.
The \code{event_age} option aggregates by the group of the age at the event timing.
\code{event_age} requires the \code{bname} to be provided in the model.}

\item{na.rm}{Logical. If \code{TRUE}, remove \code{NA} values for the aggregation. The default is \code{TRUE}.}
}
\value{
A \code{tibble} with the summary statistics
}
\description{
Aggregate the mean and variance of the estimated individual-level child penalties
}
