% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_unitdid}
\alias{aggregate_unitdid}
\title{Aggregate the mean and variance of the estimated unit-level DiD effects}
\usage{
aggregate_unitdid(
  object,
  agg = "full",
  na.rm = TRUE,
  by = NULL,
  normalized = NULL,
  only_full_horizon = TRUE
)
}
\arguments{
\item{object}{\code{unitdid} object}

\item{agg}{Aggregation method.
One of \code{c("full", "event", "event_age")} and the default is \code{full}.
If \code{by} is provided in the model,
all the options will separately aggregate by its group.
The \code{event} option aggregates by the group of the event timing.
The \code{event_age} option aggregates by the group of the age at the event time.
\code{event_age} requires the \code{bname} to be provided in the model.}

\item{na.rm}{Logical. If \code{TRUE}, remove \code{NA} values for the aggregation.
The default is \code{TRUE}.}

\item{by}{A character vector of variables to aggregate separately by.
Default is inherited from the \code{unitdid} object but you can override it here.
You can estimate the unit-level DiD effects separately by \code{by} in \code{unitdid}
but you can also aggregate the estimates by (higher-level) \code{by} here.
You can use "rel_time" as the highest level of aggregation.}

\item{normalized}{Logical. If \code{TRUE}, the function will normalize
the aggregated mean and variance by the mean of the imputed outcome variable.
Default is inherited from the \code{unitdid} object.}

\item{only_full_horizon}{Logical. If TRUE, when you aggregate
the unit-level treatment effect, only the event year (\code{ename})
with full horizon (\code{k_min:k_max}) will be included.
This is recommended in the case that you do not want to change
the composition of the event year (or age for the child penalties)
for each estimated point in \code{k_min:k_max}. Default is TRUE.}
}
\value{
A \code{tibble} with the aggregated mean and variance of
the estimated unit-level DiD effects
}
\description{
Aggregate the mean and variance of the estimated unit-level DiD effects
}
