% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\docType{data}
\name{base_indcp}
\alias{base_indcp}
\title{Simulated Individual Child Panalty Data}
\format{
\subsection{\code{base_indcp}}{

A dataframe with 1000 individuals for each birth year from 1965 to 1984:
\describe{
\item{\code{id}}{Individual identifier}
\item{\code{year}}{Year of observation}
\item{\code{byear}}{Birth year}
\item{\code{cage}}{Age at first birth}
\item{\code{rel_time}}{Relative time to first birth}
\item{\code{y}}{Outcome variable}
}
}
}
\source{
Generated by \code{gen_data()} with seed \code{1234}
}
\usage{
base_indcp
}
\description{
Simulated Individual Child Panalty Data
}
\keyword{datasets}
