% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indcp.R
\name{est_indcp}
\alias{est_indcp}
\title{Title}
\usage{
est_indcp(data, yname, iname, tname, bname, treatment, weights = NULL)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{iname}{Individual identifier}

\item{tname}{Time variable}

\item{bname}{Birth year variable}

\item{treatment}{Indicator variable for the treatment}

\item{weights}{Optional. Variable name for regression weights}
}
\value{
dataframe with the residualized outcome variable
}
\description{
Title
}
\examples{
library(indcp)
est_indcp(yname = "y", iname = "id", tname = "year", bname = "byear",
          treatment = "is_treated", data = base_indcp)

}
