% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_heterocp}
\alias{gen_heterocp}
\title{Generate Sample Heterogenous Child Penalty Data}
\usage{
gen_heterocp(size_cohort = 300)
}
\arguments{
\item{size_cohort}{n_obsumber of individuals per birth year}
}
\value{
A sample dataframe with heterogenous child penalty over the age at first birth
}
\description{
Generate Sample Heterogenous Child Penalty Data
}
\examples{
set.seed(1234)
base_heterocp <- gen_heterocp()

}
