% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indcp.R
\name{get_indcp}
\alias{get_indcp}
\title{Get individual-level child penalties}
\usage{
get_indcp(object)
}
\arguments{
\item{object}{\code{indcp} object}
}
\value{
A dataframe with estimated individual-level child penalties
}
\description{
Get individual-level child penalties
}
