% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indcp.R
\name{indcp}
\alias{indcp}
\title{A function estimates individual-level child penalties}
\usage{
indcp(
  data,
  yname,
  iname,
  tname,
  bname,
  kname,
  aname = "cage",
  k_min = 0,
  k_max = 5,
  compute_var_me = FALSE,
  only_full_horizon = TRUE
)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{iname}{Individual identifier}

\item{tname}{Time variable}

\item{bname}{Birth year variable}

\item{kname}{Relative time to treatment variable}

\item{aname}{Optional. Age at treatment variable. If not provided, it will be computed as t - b - k and added to the dataframe as \code{cage}.}

\item{k_min}{Relative time to treatment at which treatment starts. Default is 0.}

\item{k_max}{Relative time to treatment at which treatment ends. Default is 5.}

\item{compute_var_me}{Logical. If TRUE, the function will compute the variance of the measurement errors and the variance of the individual-level child-penalties. Default is FALSE.}

\item{only_full_horizon}{Logical. If TRUE, when you aggregate the individual child penalties, only the cohorts with full horizon (\code{k_min:k_max}) will be included. Default is TRUE.}
}
\value{
A \code{indcp} class object.
}
\description{
A function estimates individual-level child penalties
}
