% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.indcp}
\alias{summary.indcp}
\title{Title}
\usage{
\method{summary}{indcp}(object, agg = "full", compute_var = FALSE)
}
\arguments{
\item{object}{\code{indcp} object}

\item{agg}{Aggregation method. Default is \code{full}.}

\item{compute_var}{Logical. If \code{TRUE}, compute variance of the estimated.}
}
\value{
dataframe with the summary statistics
}
\description{
Title
}
