% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.unitdid}
\alias{summary.unitdid}
\title{Aggregate the mean and variance of the estimated unit-level DiD effects}
\usage{
\method{summary}{unitdid}(object, ...)
}
\arguments{
\item{object}{\code{unitdid} object}

\item{...}{\code{aggregate_unitdid} arguments}
}
\value{
A \code{tibble} with the summary statistics
}
\description{
Aggregate the mean and variance of the estimated unit-level DiD effects
}
\examples{
library(unitdid)
mdl_base <- base_heterocp |>
  unitdid(yname = "y",
        iname = "id",
        tname = "year",
        ename = "cyear",
        bname = "byear")
summary(mdl_base, agg = "event_age")

}
