% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitdid-package.R
\docType{package}
\name{unitdid-package}
\alias{unitdid-package}
\title{unitdid: Unit-level Difference-in-Difference Estimator}
\description{
The package provides a function to estimate the unit-level difference-in-difference estimator proposed by Arkhangelsky, Yanagimoto, and Zohar (2024).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kazuyanagimoto.com/unitdid/}
  \item \url{https://github.com/kazuyanagimoto/unitdid/}
}

}
\author{
\strong{Maintainer}: Kazuharu Yanagimoto \email{kazuharu.yanagimoto@gmail.com} (\href{https://orcid.org/0009-0007-1967-8304}{ORCID})

}
\keyword{internal}
