% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_indcp}
\alias{aggregate_indcp}
\title{Aggregate the mean and variance of the estimated individual-level child penalties}
\usage{
aggregate_indcp(object, agg = "full")
}
\arguments{
\item{object}{\code{indcp} object}

\item{agg}{Aggregation method. One of \code{c("full", "cage", "cage_byear")} and the default is \code{full}. The \code{cage} means the age at the treatment. The \code{cage_byear} means the age at the treatment and the birth year.}
}
\value{
A \code{tibble} with the summary statistics
}
\description{
Aggregate the mean and variance of the estimated individual-level child penalties
}
