% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_unitdid}
\alias{aggregate_unitdid}
\title{Aggregate the mean and variance of the estimated unit-level DiD effects}
\usage{
aggregate_unitdid(object, agg = "full", normalized = NULL)
}
\arguments{
\item{object}{\code{unitdid} object}

\item{agg}{Aggregation method.
One of \code{c("full", "event", "event_age")} and the default is \code{full}.
If \code{by} is provided in the model,
all the options will separately aggregate by its group.
The \code{event} option aggregates by the group of the event timing.
The \code{event_age} option aggregates by the group of the age at the event timing.
\code{event_age} requires the \code{bname} to be provided in the model.}

\item{normalized}{Logical. If \code{TRUE}, the function will normalize the aggregated mean and variance
by the mean of the imputed outcome variable. Default is inherited from the \code{unitdid} object.}
}
\value{
A \code{tibble} with the aggregated mean and variance of the estimated unit-level DiD effects
}
\description{
Aggregate the mean and variance of the estimated unit-level DiD effects
}
