% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indcp.R
\name{est_indcp}
\alias{est_indcp}
\title{Title}
\usage{
est_indcp(data, yname, bname, first_stage, treatment, weights = NULL)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{bname}{Birth year variable}

\item{first_stage}{Fixed effects and other covariates to residualize in the first stage}

\item{treatment}{Indicator variable for the treatment}

\item{weights}{Optional. Variable name for regression weights}
}
\value{
dataframe with the residualized outcome variable
}
\description{
Title
}
\examples{
library(indcp)
est_indcp(yname = "y", bname = "byear", treatment = "is_treated",
         first_stage = ~ 0 | id + year, data = base_indcp)
}
