% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate Sample Data}
\usage{
gen_data(size_cohort = 1000)
}
\arguments{
\item{size_cohort}{n_obsumber of individuals per birth year}
}
\value{
A sample dataframe with heterogenous child penalty over the age at first birth
}
\description{
Generate Sample Data
}
\examples{
set.seed(1234)
base_indcp <- gen_data()

}
