% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{get_unitdid}
\alias{get_unitdid}
\title{Get unit-level Difference-in-Differences estimates}
\usage{
get_unitdid(object, normalized = NULL, export = TRUE)
}
\arguments{
\item{object}{\code{unitdid} object}

\item{normalized}{Logical. If \code{TRUE}, the function will normalize them
by the mean of the imputed outcome variable.
Default is inherited from the \code{unitdid} object.}

\item{export}{Logical. If \code{TRUE}, the function will not export the columns
with the \code{zz000} prefix, which are used in the internal computation.}
}
\value{
A dataframe with a new column of the unit-level DiD estimates
}
\description{
Get unit-level Difference-in-Differences estimates
}
