% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unitdid.R
\name{get_unitdid}
\alias{get_unitdid}
\title{Get unit-level Difference-in-Differences estimates}
\usage{
get_unitdid(object, normalized = NULL)
}
\arguments{
\item{object}{\code{unitdid} object}

\item{normalized}{Logical. If \code{TRUE}, the function will normalize them by the mean of the imputed outcome variable.
Default is inherited from the \code{unitdid} object.}
}
\value{
A dataframe with a new column of the unit-level DiD estimates
}
\description{
Get unit-level Difference-in-Differences estimates
}
