% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indcp.R
\name{indcp}
\alias{indcp}
\title{Title}
\usage{
indcp(data, yname, iname, tname, bname, kname, aname = NULL, k_min = 0)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{iname}{Individual identifier}

\item{tname}{Time variable}

\item{bname}{Birth year variable}

\item{kname}{Relative time to treatment variable}

\item{aname}{Optional. Age at treatment variable. If not provided, it will be computed as t - b - k and added to the dataframe as \code{cage}.}

\item{k_min}{Relative time to treatment at which treatment starts. Default is 0.}
}
\value{
dataframe with the residualized outcome variable
}
\description{
Title
}
