% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.indcp}
\alias{summary.indcp}
\title{Aggregate the mean and variance of the estimated individual-level child penalties}
\usage{
\method{summary}{indcp}(object, ...)
}
\arguments{
\item{object}{\code{indcp} object}

\item{...}{\code{aggregate_indcp} arguments}
}
\value{
A \code{tibble} with the summary statistics
}
\description{
Aggregate the mean and variance of the estimated individual-level child penalties
}
\examples{
library(indcp)
mdl_base <- base_indcp |>
  indcp(yname = "y",
        iname = "id",
        tname = "year",
        ename = "cyear",
        bname = "byear")
summary(mdl_base, agg = "event_age")

}
