% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.indcp}
\alias{summary.indcp}
\title{Aggregate the mean and variance of the estimated individual-level child penalties}
\usage{
\method{summary}{indcp}(object, agg = "full")
}
\arguments{
\item{object}{\code{indcp} object}

\item{agg}{Aggregation method. One of \code{c("full", "cage", "cage_byear")} and the default is \code{full}. The \code{cage} means the age at the treatment. The \code{cage_byear} means the age at the treatment and the birth year.}
}
\value{
A \code{tibble} with the summary statistics
}
\description{
Aggregate the mean and variance of the estimated individual-level child penalties
}
\examples{
library(indcp)
mdl_base <- base_indcp |>
  indcp(yname = "y",
        iname = "id",
        tname = "year",
        bname = "byear",
        kname = "rel_time")
summary(mdl_base, agg = "cage")

}
