% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.indcp}
\alias{summary.indcp}
\title{Aggregate the mean and variance of the estimated individual-level child penalties}
\usage{
\method{summary}{indcp}(object, agg = "full", compute_var = FALSE, ...)
}
\arguments{
\item{object}{\code{indcp} object}

\item{agg}{Aggregation method. Default is \code{full}.}

\item{compute_var}{Logical. If \code{TRUE}, compute variance of the estimated.}

\item{...}{Additional arguments to be passed to \code{aggregate_mean} or \code{aggregate_var}.}
}
\value{
A \code{tibble} with the summary statistics
}
\description{
Aggregate the mean and variance of the estimated individual-level child penalties
}
