% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitdid.R
\name{unitdid}
\alias{unitdid}
\title{A function estimates unit-level difference-in-differences}
\usage{
unitdid(
  data,
  yname,
  iname,
  tname,
  ename,
  first_stage = NULL,
  wname = NULL,
  k_min = 0,
  k_max = 5,
  compute_varcov = "none",
  by = NULL,
  bname = NULL,
  normalized = FALSE,
  newnames = NULL
)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{iname}{Unit identifier}

\item{tname}{Time variable}

\item{ename}{Event timing variable}

\item{first_stage}{Formula for Y(0).
Formula follows \code{\link[fixest:feols]{fixest::feols}}.
If not specified, unit (\code{iname}) and time (\code{tname}) fixed effects
will be used.}

\item{wname}{Optional. The name of the weight variable.}

\item{k_min}{Relative time to treatment at which treatment starts.
Default is 0.}

\item{k_max}{Relative time to treatment at which treatment ends.
Default is 5.}

\item{compute_varcov}{One of c("none", "var", "cov") and Default is "none".
If "var", the function will estimate the unit-level variance of the outcome
variable. If "cov", the function will estimate the unit-level covariance of
the outcome variable for each pair within \code{k_min:k_max}.}

\item{by}{A character vector of variables to estimate separately by.
Default is NULL.}

\item{bname}{Birth year variable. Default is NULL.
Necessary to aggregate the estimates by age at event.}

\item{normalized}{Logical. If TRUE,
the function will normalize the outcome variable scale. Default is FALSE.}

\item{newnames}{Optional. A list of new names for the output variables.
\code{ytildename} is the name of the imputed outcome variable.
Default is \code{paste0(yname, "_tilde")}.
\code{yvarname} is the name of the unit-level variance of the outcome variable.
Default is \code{paste0(yname, "_var")}.
\code{yvarrawname} is the name of the raw unit-level variance of
the outcome variable, which is the variance before subtracting the
variance of the measurement error. Default is \code{paste0(yname, "_varraw")}.
\code{yvarerrname} is the name of the unit-level variance of the measurement error.
Default is \code{paste0(yname, "_varerr")}.
\code{ycovname} is the name of the unit-level covariance of the outcome variable.
Default is \code{paste0(yname, "_cov")}.
\code{ycovrawname} is the name of the raw unit-level covariance of
the outcome variable, which is the covariance before subtracting the
covariance of the measurement error. Default is \code{paste0(yname, "_covraw")}.
\code{ycoverrname} is the name of the unit-level covariance of the measurement error.
Default is \code{paste0(yname, "_coverr")}.
\code{kprimename} is the name of the relative time to treatment.
This is used for the second column name of the relative time of
the unit-level covariance estimation. Default is "kprime".}
}
\value{
A \code{unitdid} class object.
}
\description{
A function estimates unit-level difference-in-differences
}
