% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitdid.R
\name{unitdid}
\alias{unitdid}
\title{A function estimates unit-level difference-in-differences}
\usage{
unitdid(
  data,
  yname,
  iname,
  tname,
  ename,
  first_stage = NULL,
  wname = NULL,
  ytildename = NULL,
  yvarname = NULL,
  k_min = 0,
  k_max = 5,
  compute_var = FALSE,
  by = NULL,
  bname = NULL,
  normalized = FALSE
)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{iname}{Unit identifier}

\item{tname}{Time variable}

\item{ename}{Event timing variable}

\item{first_stage}{Formula for Y(0).
Formula follows \code{\link[fixest:feols]{fixest::feols}}.
If not specified, unit (\code{iname}) and time (\code{tname}) fixed effects
will be used.}

\item{wname}{Optional. The name of the weight variable.}

\item{ytildename}{Optional. The name of the imputed outcome variable.
If not provided, the function will use \code{paste0(yname, "_tilde")}.}

\item{yvarname}{Optional. The name of the unit-level variance of
the outcome variable. If not provided, the function will use
\code{paste0(yname, "_var")}.}

\item{k_min}{Relative time to treatment at which treatment starts.
Default is 0.}

\item{k_max}{Relative time to treatment at which treatment ends.
Default is 5.}

\item{compute_var}{Logical. If TRUE, the function will compute
the variance of the measurement errors and the variance of the unit-level
treatment effects. Default is FALSE.}

\item{by}{A character vector of variables to estimate separately by.
Default is NULL.}

\item{bname}{Birth year variable. Default is NULL.
Necessary to aggregate the estimates by age at event.}

\item{normalized}{Logical. If TRUE,
the function will normalize the outcome variable scale. Default is FALSE.}
}
\value{
A \code{unitdid} class object.
}
\description{
A function estimates unit-level difference-in-differences
}
